Fixed so that changes to padding is detected when broadcasting "elastic:adjust".
The textarea's padding was only read during the initialization of the
directive. If the padding changed during the lifetime of element, then
the new padding was not taken into account, leading to incorrect (too
low) height being calculated.

These calculations have now been moved into a separate function, which
is called during initialization and when an "elastic:adjust" event is
received. That at least gives the developer a way to change padding,
even if it requires broadcasting the "elastic:adjust" event after the
change.