Code refactored, to enable changing border and height of the textarea, when directive used in Modal Window.
When the directive is used in a Modal Window, angular-material md-dialog or ui-bootstrap modal dialog, the textarea is already present without a value, due to which the height, border calculated are incorrect, furthermore, on updating the ng-model, i.e. when the user starts typing, the textarea behaves inappropriately. To solve this, the code responsible for calculating the textarea height and border is moved into a function which then can be called when an event 'elastic:recalculate' event is fired.